pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aeroshape"
version = "3.0.0"
description = "NURBS-based geometry engine for aircraft design and analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Victor Alulema"},
    {name = "Esteban Valencia"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Intended Audience :: Science/Research",
]

dependencies = [
    "numpy>=1.20",
    "cadquery-ocp>=7.7",
]

[project.optional-dependencies]
gui = [
    "streamlit>=1.20",
    "plotly>=5.0",
    "pandas>=1.3",
]
viz = [
    "vedo>=2024.5",
    "matplotlib>=3.5",
]
all = [
    "aeroshape[gui,viz]",
]

[project.urls]
Repository = "https://github.com/victor-alulema/aeroshape"

[tool.setuptools.packages.find]
include = ["aeroshape*"]
